A lot of changes; code remodeling, adding a test-server * removed the F1Stream object * added ByteBuffer object to improve raw data handling of the stream reader * removed globalvar.py module now there are no global variables needed any more * renamed f1*.py to ./src/*.py
capture application added; this only read the streams (http and raw TCP 4321) and stores the data into files
in the ./keyframe/X<eventid>/ folder to be used with test/replay server.

test/replay server added; this replays the captured data, this is not yet done. Currently the http part works
and listen and accepting the raw TCP sockets, data reading and sending not yet finished.